feat: support ip V

+++ scripts/download-ripe-data.sh
@@ -1,13 +1,35 @@
 #!/bin/sh
 
-
 url="https://ftp.ripe.net/ripe/dbase/split/ripe.db.inetnum.gz"
+ipV6Url="https://ftp.ripe.net/ripe/dbase/split/ripe.db.inet6num.gz"
 destination="db"
-echo "Downloading file from $url..."
-wget -q "$url" -P "$destination"
+force="false"
 
-echo "Extracting file..."
-gzip -d "$destination/ripe.db.inetnum.gz"
+while [ $# -gt 0 ]; do
+    if [ "$1" = "--force" ]; then
+        force="true"
+    fi
+    shift
+done
 
-rm -rf "$destination/ripe.db.inetnum.gz"
-echo "Extraction complete."
+if [ ! -f "$destination/ripe.db.inetnum" ] || [ "$force" = "true" ]; then
+    echo "Downloading file from $url..."
+    wget -q "$url" -P "$destination"
+    echo "Extracting file..."
+    gzip -d "$destination/ripe.db.inetnum.gz"
+    echo "Extraction complete."
+    rm -rf "$destination/ripe.db.inetnum.gz"
+else
+    echo "File already exists in $destination. Use --force to download again."
+fi
+
+if [ ! -f "$destination/ripe.db.inet6num" ] || [ "$force" = "true" ]; then
+    echo "Downloading IPv6 file from $ipV6Url..."
+    wget -q "$ipV6Url" -P "$destination"
+    echo "Extracting file..."
+    gzip -d "$destination/ripe.db.inet6num.gz"
+    echo "Extraction complete."
+    rm -rf "$destination/ripe.db.inet6num.gz"
+else
+    echo "IPv6 file already exists in $destination. Use --force to download again."
+fi
